requirements.txt
PyMySQL==1.1.0
pytest==7.2.1
PyYAML==6.0.1
